requirements.txt
fastapi
uvicorn
pymongo
requests
pyyaml